move new tests to test_99 file to avoid conflicts
